+++ README.md
@@ -1,4 +1,4 @@
-# 住所録第4版
+# 住所録第5版
 
-住所録を管理するためのプログラムの第4版です。
-
+住所録を管理するためのプログラムの第5版です。
+モジュール化を実施し、表示系の関数群を別モジュールとして切り出しました。
